Move platform dependent function
